+++ CHANGES.md
@@ -53,10 +53,16 @@ with Reader("tests/fixtures/cog_tags.tif") as src:
 ```
 
 * only cast data to `uint8` if colormap values are of type `uint8`
+
 * add `alpha_mask` attribute to `ImageData` class
+
 * allow partial alpha values from alpha band
+
 * better handle non-uint8 alpha band
+
 * remove deprecated `force_binary_mask` option in `reader.read` function  **breaking change**
+
+* add `nodata`, `scales` and `offsets` attributes to `ImageData` and `PointData` classes
 
 # 7.8.1 (2025-06-16)